chore: chain_performance_by_mq: remove `chain/types` dependency
There is a strange bug cause this tool fail to compile during static
linking to snappy in `release` compilation.

Since we only use a constant from `cita/types` to do tests. We can
directly hardcode this constant and write a test to check whether
it's changed in `cita/types`.

This should temporarily work around compilation failure.